test(wallet): gather rotation scenarios into one suite
Move the rebind-on-refresh tests (wallet-init) and receive-across-a-rotation
tests (wallet-receive) into a new wallet-rotation.node.test.ts, along with the
rotation fixtures they duplicated. Tests moved verbatim: same names, same
assertions, same msw handlers. Gives future rotation coverage an obvious home.
